Activate beartype claw on lcm.solution and lcm.simulation
Extends the AST-rewriting claw to the post-perimeter simulation and
solution packages with a plain `INTERNAL_CONF` (violations surface as
`BeartypeCallHintViolation` — an internal-bug signal, distinct from the
perimeter packages' project-exception confs).

Drift surfaced and fixed:
- `transitions.random_id`: vmap body returns a 0-d scalar; return hint
  corrected `Int1D` -> `ScalarInt`.
- `result._extract_period_data`: `subject_id` built without a pinned
  dtype; pinned to `int32`.
- `initial_conditions._check_regime_feasibility`: `idx_arr` built from a
  Python list without a pinned dtype; pinned to `int32`.

Canonicalization barrier:
- New `canonicalize_initial_conditions` casts every initial-conditions
  array to its canonical pylcm dtype at the `Model.simulate` perimeter,
  before validation and the simulate stack. Downstream code keeps strict
  canonical annotations.

Tests:
- `test_beartype_claw.py`: claw-liveness checks for both packages.
- Direct-call tests updated to pass canonical-dtype inputs (the claw now
  enforces what these internal functions always required); `InternalRegimeMock`
  subclasses `InternalRegime` so `isinstance` holds at solve()'s boundary.

+++ src/lcm/simulation/result.py
@@ -527,7 +527,7 @@ def _extract_period_data(
 ) -> dict[str, FloatND | IntND | BoolND]:
     """Extract data from a single period's simulation results."""
     data: dict[str, FloatND | IntND | BoolND] = {
-        "subject_id": jnp.arange(len(result.in_regime)),
+        "subject_id": jnp.arange(len(result.in_regime), dtype=jnp.int32),
         "period": jnp.full_like(result.in_regime, period, dtype=jnp.int32),
         "_in_regime": result.in_regime,
         "value": result.V_arr,
